完成用户登录认证 电影选座功能

+++ service/routes/api/UserRouter.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const UserService = require('../../service/UserService');
 const { asyncHandler, sendMsg } = require('../util/util');
-
+const jwt = require('../util/jwt');
 const router = express.Router();
 
 /**
@@ -9,6 +9,11 @@ const router = express.Router();
  */
 router.post('/login', asyncHandler(async (req, res) => {
     const result = await UserService.loginUser(req.body.loginId, req.body.loginPwd);
+    if (result) {
+        jwt.publish(res, undefined, {
+            UserId: result.id.toString()
+        })
+    }
     return result;
 }));
 
@@ -50,14 +55,33 @@ router.post('/', asyncHandler(async (req, res) => {
  * 修改用户信息
  */
 router.put('/:id', asyncHandler(async (req, res) => {
-    return await UserService.updateUser(req.params.id, req.body);
+    if (req.token.UserId || req.token.Admin) {
+        return await UserService.updateUser(req.params.id, req.body);
+    } else {
+        res.status(403).send(sendMsg(403, "没有权限访问该接口！"));
+    }
 }));
 
 /**
  * 通过ID删除用户
  */
 router.delete('/:id', asyncHandler(async (req, res) => {
-    return await UserService.deleteUser(req.params.id);
+    if (req.token.UserId || req.token.Admin) {
+        return await UserService.deleteUser(req.params.id);
+    } else {
+        res.status(403).send(sendMsg(403, "没有权限访问该接口！"));
+    }
+}));
+
+/**
+ * 获取用户身份
+ */
+router.post('/whoami', asyncHandler(async (req, res) => {
+    if (req.token.UserId) {
+        return await UserService.getUserFindById(req.token.UserId);
+    } else {
+        res.send(null);
+    }
 }));
 
 module.exports = router;
